Tweak grenade physics, rocket sound IDs; rm rocket smoke.

+++ src/tasks/rocket.h
@@ -5,10 +5,10 @@ struct taskRocket {
 	offset accel;
 	bool dead, live;
 	uint16_t ttl;
-	uint32_t soundId;
+	uint16_t soundId;
 };
 #define rocketFromMover(x) ((taskRocket*)((char*)(x) - offsetof(taskRocket, m)))
 
 extern void taskRocket_draw(void *_data);
-extern void taskRocket_create(gamestate *gs, offset p1, offset vel, unitvec dir, box *parent, uint32_t soundId);
+extern void taskRocket_create(gamestate *gs, offset p1, offset vel, unitvec dir, box *parent, u16 soundId);
 extern void taskRocket_define(taskDefn *t);
